hiltend-backend/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
annotated-doc==0.0.4
    # via
    #   fastapi
    #   typer
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
azure-core==1.39.0
    # via
    #   azure-identity
    #   azure-storage-blob
    #   azure-storage-file-datalake
azure-identity==1.25.3
    # via hiltend-backend (pyproject.toml)
azure-storage-blob==12.28.0
    # via azure-storage-file-datalake
azure-storage-file-datalake==12.23.0
    # via hiltend-backend (pyproject.toml)
certifi==2026.2.25
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.7
    # via requests
click==8.3.2
    # via
    #   rich-toolkit
    #   typer
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
cryptography==46.0.7
    # via
    #   azure-identity
    #   azure-storage-blob
    #   msal
    #   pyjwt
dnspython==2.8.0
    # via email-validator
email-validator==2.3.0
    # via
    #   fastapi
    #   pydantic
fastapi==0.136.0
    # via hiltend-backend (pyproject.toml)
fastapi-cli==0.0.24
    # via fastapi
fastapi-cloud-cli==0.17.0
    # via fastapi-cli
fastar==0.11.0
    # via
    #   fastapi
    #   fastapi-cloud-cli
greenlet==3.4.0
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
idna==3.11
    # via
    #   anyio
    #   email-validator
    #   httpx
    #   requests
isodate==0.7.2
    # via
    #   azure-storage-blob
    #   azure-storage-file-datalake
jinja2==3.1.6
    # via fastapi
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
msal==1.36.0
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.3.1
    # via azure-identity
pycparser==3.0
    # via cffi
pydantic==2.13.1
    # via
    #   fastapi
    #   fastapi-cloud-cli
    #   pydantic-extra-types
    #   pydantic-settings
pydantic-core==2.46.1
    # via pydantic
pydantic-extra-types==2.11.1
    # via fastapi
pydantic-settings==2.13.1
    # via fastapi
pygments==2.20.0
    # via rich
pyjwt==2.12.1
    # via msal
pyodbc==5.3.0
    # via hiltend-backend (pyproject.toml)
python-dotenv==1.2.2
    # via
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.26
    # via fastapi
pyyaml==6.0.3
    # via uvicorn
requests==2.33.1
    # via
    #   azure-core
    #   msal
rich==15.0.0
    # via
    #   rich-toolkit
    #   typer
rich-toolkit==0.19.7
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
rignore==0.7.6
    # via fastapi-cloud-cli
sentry-sdk==2.58.0
    # via fastapi-cloud-cli
shellingham==1.5.4
    # via typer
sqlalchemy==2.0.49
    # via hiltend-backend (pyproject.toml)
starlette==1.0.0
    # via fastapi
typer==0.24.1
    # via
    #   fastapi-cli
    #   fastapi-cloud-cli
typing-extensions==4.15.0
    # via
    #   azure-core
    #   azure-identity
    #   azure-storage-blob
    #   azure-storage-file-datalake
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   pydantic-extra-types
    #   rich-toolkit
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   requests
    #   sentry-sdk
uvicorn==0.44.0
    # via
    #   fastapi
    #   fastapi-cli
    #   fastapi-cloud-cli
watchfiles==1.1.1
    # via uvicorn
websockets==16.0
    # via uvicorn
